Harden TypeScript alias mutation coverage
