+++ packages/plugin-typescript/src/aliasImport/edgeType.ts
@@ -1,0 +1,8 @@
+export const TYPESCRIPT_ALIAS_IMPORT_EDGE_TYPE = {
+  id: 'codegraphy.typescript:alias-import',
+  label: 'TypeScript Alias Import',
+  defaultColor: '#38BDF8',
+  defaultVisible: true,
+} as const;
+
+export const COMPILER_OPTIONS_PATHS_SOURCE_ID = 'compiler-options-paths';
